+++ crates/autopilot/src/boundary/mod.rs
@@ -1,15 +1,12 @@
-use {crate::domain, ethrpc::Web3, std::collections::HashMap, url::Url};
 pub use {
-    crate::{
-        database::{
-            competition::Competition,
-            order_events::{store_order_events, OrderEventLabel},
-        },
-        driver_model::{reveal, settle, solve},
+    crate::database::{
+        competition::Competition,
+        order_events::{store_order_events, OrderEventLabel},
     },
     database,
     model::{
         app_data::AppDataHash,
+        bytes_hex,
         interaction::InteractionData,
         order::{
             BuyTokenDestination,
@@ -26,6 +23,7 @@ pub use {
     },
     shared::order_validation::{is_order_outside_market_price, Amounts},
 };
+use {crate::domain, ethrpc::Web3, std::collections::HashMap, url::Url};
 
 pub mod events;
 pub mod order;
