+++ crates/autopilot/src/infra/solvers/dto/mod.rs
@@ -1,0 +1,7 @@
+//! Types for communicating with drivers as defined in
+//! `crates/driver/openapi.yml`.
+
+pub mod quote;
+pub mod reveal;
+pub mod settle;
+pub mod solve;
