+++ crates/autopilot/src/lib.rs
@@ -3,7 +3,6 @@ pub mod boundary;
 pub mod database;
 pub mod decoded_settlement;
 pub mod domain;
-pub mod driver_model;
 pub mod event_updater;
 pub mod infra;
 pub mod on_settlement_event_updater;
